% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS.R
\docType{data}
\name{LOS}
\alias{LOS}
\title{Length of hospital stays of 32 patients.}
\format{A vector of 32 integers
\describe{
  \item{LOS}{length of hospital stay, in days}
}}
\source{
Marazzi, A. and Yohai, V. J. (2010). Optimal robust estimates using the Hellinger distance. Advances in data analysis and classification, 4(2-3), 169-179.
}
\usage{
LOS
}
\description{
A dataset containing the length of hospital stays of 32 patients.
}
\keyword{datasets}
