% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{psi_binom}
\alias{psi_binom}
\title{computes the psi function of the binomial MI-estimator}
\usage{
psi_binom(y, n, p)
}
\arguments{
\item{y}{An integer in [0,n]}

\item{n}{A positive integer}

\item{p}{A real number in (0,1)}
}
\value{
A real number in [0,1]
}
\description{
computes the psi function of the binomial MI-estimator
}
