% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{psi_binom_s}
\alias{psi_binom_s}
\title{computes the score function of the binomial MI-estimator}
\usage{
psi_binom_s(mstr, n, p)
}
\arguments{
\item{mstr}{A  vector of integers in [0,n]}

\item{n}{A positive integer}

\item{p}{A real number in (0,1)}
}
\value{
A real number
}
\description{
computes the score function of the binomial MI-estimator
}
