% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{psi_exp}
\alias{psi_exp}
\title{computes the psi function of the exponential MI-estimator}
\usage{
psi_exp(lambda, y)
}
\arguments{
\item{lambda}{A positive real number}

\item{y}{A non negative real number}
}
\value{
A real number in [0,1]
}
\description{
computes the psi function of the exponential MI-estimator
}
