% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{psi_exp_s}
\alias{psi_exp_s}
\title{computes the score function of the exponential MI-estimator}
\usage{
psi_exp_s(lambda, mstr)
}
\arguments{
\item{lambda}{A positive real number}

\item{mstr}{A vector of non negative real numbers}
}
\value{
A real number
}
\description{
computes the score function of the exponential MI-estimator
}
