% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{psi_geom_s}
\alias{psi_geom_s}
\title{computes the score function of the geometric MI-estimator}
\usage{
psi_geom_s(mu, mstr)
}
\arguments{
\item{mu}{A positive real number}

\item{mstr}{A vector of non negative integers}
}
\value{
A real number
}
\description{
computes the score function of the geometric MI-estimator
}
