% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{psi_pois}
\alias{psi_pois}
\title{computes the psi function of the poisson MI-estimator}
\usage{
psi_pois(mu, y)
}
\arguments{
\item{mu}{A positive real number}

\item{y}{A non negative integer}
}
\value{
A real number in [0,1]
}
\description{
computes the psi function of the poisson MI-estimator
}
