% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_for_negative_binomial.R
\name{psibinneg1}
\alias{psibinneg1}
\title{Computes the first entry of the score function of an MI-estimator for negative binomial distribution.}
\usage{
psibinneg1(mu, alpha, x)
}
\arguments{
\item{mu}{A possitive real number}

\item{alpha}{A possitive real number}

\item{x}{A vector of univariate observations}
}
\value{
A real number
}
\description{
Computes the first entry of the score function of an MI-estimator for negative binomial distribution.
}
\examples{
  mu <- 2
  alpha <- 1
  x <- rnbinom(100,mu=mu,size=1/alpha)
  mean(robustr:::psibinneg1(mu,alpha,x))
}
