% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{rob_estimate_binom}
\alias{rob_estimate_binom}
\title{Computes a robust estimate of the probability of success of a binomial random variable Y with known n}
\usage{
rob_estimate_binom(x, n)
}
\arguments{
\item{x}{A vector of univariate observations}

\item{n}{An positive integer}
}
\value{
The estimated parameter of the Poisson distribution
}
\description{
Computes a robust estimate of the probability of success of a binomial random variable Y with known n
}
\examples{
  n<-3
  x <- rbinom(20,n,0.1)
  rob_estimate_binom(x,n)
  mean(x)/n
}
