% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{rob_estimate_exp}
\alias{rob_estimate_exp}
\title{Computes a robust estimate of the parameter of an Exponential random variable.}
\usage{
rob_estimate_exp(x)
}
\arguments{
\item{x}{A vector of univariate observations}
}
\value{
The estimated parameter of the exponential distribution
}
\description{
Computes a robust estimate of the parameter of an Exponential random variable.
}
\examples{
  x <- rexp(50,2)
  rob_estimate_exp(x)
  MASS::fitdistr(x,dens="exponential")
}
