% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_for_one_parameter_distributions.R
\name{rob_estimate_geom}
\alias{rob_estimate_geom}
\title{Computes a robust estimate of the parameter of a geometric random variable Y}
\usage{
rob_estimate_geom(x)
}
\arguments{
\item{x}{A vector of univariate observations}
}
\value{
The estimated parameter of the Poisson distribution
}
\description{
Computes a robust estimate of the parameter of a geometric random variable Y
}
\examples{
  x <- rgeom(50,1/2)
  rob_estimate_geom(x)
  MASS::fitdistr(x,dens="geometric")
}
