% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_for_negative_binomial.R
\name{rob_estimate_negbin}
\alias{rob_estimate_negbin}
\title{Computes a robust estimate of the parameters of a negative binomial random variable.}
\usage{
rob_estimate_negbin(x, start = NULL)
}
\arguments{
\item{x}{A vector of univariate observations}

\item{start}{Initial vector of parameters to start the optimization}
}
\value{
The estimated vector of parameters. The first entry is the mean and the second is the dispersion parameter, that equals 1/size.
}
\description{
Computes a robust estimate of the parameters of a negative binomial random variable.
}
\examples{
  library(MASS)
  x <- rnbinom(100,mu=2,size=1)
  MASS::fitdistr(x,dens="negative binomial")
  rob_estimate_negbin(x)
}
