% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_for_negative_binomial.R
\name{score_mi_negbin}
\alias{score_mi_negbin}
\title{Computes the score function of a MI-estimator for negative binomial distribution.}
\usage{
score_mi_negbin(par, x)
}
\arguments{
\item{par}{A two dimensional vector of parameters for negative binomial distribution. The first entry is the mean and the second is the dispertion parameter, that equals 1/size.}

\item{x}{A vector of univariate observations}
}
\value{
A real number
}
\description{
Computes the score function of a MI-estimator for negative binomial distribution.
}
